+++ plugin/blog/blog.js
@@ -29,10 +29,12 @@ function get(id, n) {
       $('#news-' + n + ' > header')  .html('<h2>' + data['title'] + '</h2>');
       $('#news-' + n + ' > header').append('<p>News scritta da <em>' + data['author'] + '</em> il <em>' + data['date'] + '</em>. <a class="roll" href="id/' + id + '.html"><span data-title="Commenti">Commenti</span></a><br /><em>' + $.map(data['tags'].split(', '), function(tag) { return '<a class="roll" href="tag/' + tag + '"><span data-title="' + tag + '">' + tag + '</span></a>' }) +'</em></p>');
       $('#news-' + n)              .append('<article>' + data['text'].replace(/\.(png|jpg|gif)\"/g, '.$1" data-lightbox="roadtrip"') + '</article>');
+      
       if(screen.height < 1080 && screen.height >= 1024) {
         $('#news-' + n).prepend('<div class="spacing"></div>');
         $('#news-' + n).append('<div class="spacing"></div>');
       }
+
       if(screen.height < 1024 ) {
         $('#news-' + n).prepend('<div class="spacing"></div>');
         $('#news-' + n).append('<div class="spacing"></div>');
@@ -42,44 +44,30 @@ function get(id, n) {
       $('title')                     .html('Error - ' + pagetitle);
       $('#news-' + n + ' > header')  .html('<h1>Error</h1>');
       $('#news-' + n)              .append('<article><h3>The article you are searching has been eaten by Shinbo.</h3><img src="css/theme/error.gif" alt="Error" /></article>');
+      
+      if(screen.height < 1080 && screen.height >= 1024) {
+        $('#news-' + n).prepend('<div class="spacing"></div>');
+        $('#news-' + n).append('<div class="spacing"></div>');
+      }
+
+      if(screen.height < 1024 ) {
+        $('#news-' + n).prepend('<div class="spacing"></div>');
+        $('#news-' + n).append('<div class="spacing"></div>');
+      }
     }
   });
 }
 
 function get_by_id(id) {
-  $.ajax({
-    type:     'GET',
-    dataType: 'json',
-    url:      'news/' + id + '.json',
-    async:    false,
-    success:  function(data) {
-      $('title')              .html(data['title'] + ' - ' + pagetitle);
-      $('#news-1 > header')   .html('<h2>' + data['title'] + '</h2>');
-      $('#news-1 > header') .append('<p>News scritta da <em>' + data['author'] + '</em> il <em>' + data['date'] + '</em>.<br /><em>' + $.map(data['tags'].split(', '), function(tag) { return '<a href="tag/' + tag + '">' + tag + '</a>' }) + '</em></p>');
-      $('#news-1')          .append('<article>' + data['text'].replace(/\.(png|jpg|gif)\"/g, '.$1" data-lightbox="roadtrip"') + '<div align="center"><script>var idcomments_acct = \'87b04b2a52c611244adf65de705ad6f4\'; var idcomments_post_id; var idcomments_post_url;</script><span id="IDCommentsPostTitle" style="display:none"></span><script type="text/javascript" src="http://www.intensedebate.com/js/genericCommentWrapperV2.js"></script></div></article>');
-      if(screen.height < 1080 && screen.height >= 1024) {
-        $('#news-1').prepend('<div class="spacing"></div>');
-        $('#news-1').append('<div class="spacing"></div>');
-      }
-      if(screen.height < 1024 ) {
-        $('#news-1').prepend('<div class="spacing"></div>');
-        $('#news-1').append('<div class="spacing"></div>');
-      }
-    },
-    error: function(data) {
-      $('title')             .html('Error - ' + pagetitle);
-      $('#news-1 > header')  .html('<h1>Error</h1>');
-      $('#news-1')           .append('<article><h3>The article you are searching has been eaten by Shinbo.</h3><img src="css/theme/error.gif" alt="Error" /></article>');
-      if(screen.height < 1080 && screen.height >= 1024) {
-        $('#news-1').prepend('<div class="spacing"></div>');
-        $('#news-1').append('<div class="spacing"></div>');
-      }
-      if(screen.height < 1024 ) {
-        $('#news-1').prepend('<div class="spacing"></div>');
-        $('#news-1').append('<div class="spacing"></div>');
-      }
-    }
-  });
+  if(screen.height < 1080 && screen.height >= 1024) {
+    $('#news-1').prepend('<div class="spacing"></div>');
+    $('#news-1').append('<div class="spacing"></div>');
+  }
+
+  if(screen.height < 1024 ) {
+    $('#news-1').prepend('<div class="spacing"></div>');
+    $('#news-1').append('<div class="spacing"></div>');
+  }
 }
 
 $(document).ready(function() {